Fix missing supports
test_swagger.py monkey patches SwaggerParser::check_type, as it does not
like `supports` field being of raw object type. There seems to be
nothing wrong with that according to swagger spec.
RESTPLUS_MASK_SWAGGER ('X-Fields' header for masking) is turned off,
otherwise the swagger_tester sends 'X-Fields: string' with the request
and the response is empty object (we'd need to somehow set an example
`*` to get non masked response)